докусентация, кодировка и тесты

+++ Matrix/main.cpp
@@ -4,14 +4,18 @@
 #include "../Solver/RandomGenerator.h"
 #include "../Solver/Exercise.h"
 
+/**
+* @brief Точка входа в программу
+* @return 0 программа выполнено успешно
+*/
 int main()
 {
     using namespace miit::data::generators;
     using namespace miit::matrix::exercises;
 
     
-    const size_t rows = 5;
-    const size_t columns = 5;
+    const int rows = 5;
+    const int columns = 5;
 
     
     auto randomGen = std::make_unique<RandomGenerator<int>>(-10, 10);
